+++ composeApp/src/iosMain/kotlin/org/onion/agro/io/SystemFileSystem.ios.kt
@@ -1,0 +1,7 @@
+package org.onion.agro.io
+
+import okio.FileSystem
+import okio.SYSTEM
+
+internal actual val systemFileSystem: FileSystem
+    get() = FileSystem.SYSTEM